перенес showErrorMsg и showSuccessMsg в модуль utils.js
