+++ js/main.js
@@ -3,7 +3,7 @@ import { createMarkers } from './map.js';
 import { setOnMapLoad } from './map.js';
 import { getData } from './fetch.js';
 import { setOnFormSubmit } from './form.js';
-import { showSuccessMsg } from './popup.js';
+import { showSuccessMsg } from './utils.js';
 import './form.js';
 
 setPageEnabled(false);
