+++ js/popup.js
@@ -2,8 +2,8 @@ import { isEscEvent } from './utils.js';
 import { resetForm } from './page.js';
 
 const cardTemplate = document.querySelector('#card').content.querySelector('.popup');
-const SuccessMsgTemplate = document.querySelector('#success').content.querySelector('.success');
-const ErrorMsgTemplate = document.querySelector('#error').content.querySelector('.error');
+const successMsgTemplate = document.querySelector('#success').content.querySelector('.success');
+const errorMsgTemplate = document.querySelector('#error').content.querySelector('.error');
 
 const getPopupTypeName = (type) => {
   switch (type) {
@@ -65,22 +65,6 @@ const showPopup = (advert) => {
   return popuptItem;
 };
 
-const showAlert = (message) => {
-  const alertContainer = document.createElement('div');
-  alertContainer.style.position = 'absolute';
-  alertContainer.style.left = '350px';
-  alertContainer.style.right = '350px';
-  alertContainer.style.top = '15px';
-  alertContainer.style.padding = '20px 3px';
-  alertContainer.style.fontSize = '25px';
-  alertContainer.style.color = 'black';
-  alertContainer.style.textAlign = 'center';
-  alertContainer.style.backgroundColor = 'red';
-
-  alertContainer.textContent = message;
-  document.body.appendChild(alertContainer);
-};
-
 const hideMessage = (message, isSuccess) => {
   message.remove();
   if (isSuccess) {
@@ -101,19 +85,9 @@ const showMessage = (message, isSuccess) => {
   });
 };
 
-const showSuccessMsg = () => {
-  const message = SuccessMsgTemplate.cloneNode(true);
-  showMessage(message, true);
-};
-
-const showErrorMsg = () => {
-  const message = ErrorMsgTemplate.cloneNode(true);
-  showMessage(message, false);
-};
-
 export {
+  successMsgTemplate,
+  errorMsgTemplate,
   showPopup,
-  showSuccessMsg,
-  showErrorMsg,
-  showAlert
+  showMessage
 };
