+++ js/utils.js
@@ -1,3 +1,5 @@
+import { successMsgTemplate, errorMsgTemplate, showMessage } from './popup.js';
+
 const getRandomFloat = (min, max, precision) => {
   if (max <= min || max < 0 || min < 0) {
     return null;
@@ -29,6 +31,32 @@ const setFormEnabled = (form, elements, enabled) => {
   }
 };
 
+const showAlert = (message) => {
+  const alertContainer = document.createElement('div');
+  alertContainer.style.position = 'absolute';
+  alertContainer.style.left = '350px';
+  alertContainer.style.right = '350px';
+  alertContainer.style.top = '15px';
+  alertContainer.style.padding = '20px 3px';
+  alertContainer.style.fontSize = '25px';
+  alertContainer.style.color = 'black';
+  alertContainer.style.textAlign = 'center';
+  alertContainer.style.backgroundColor = 'red';
+
+  alertContainer.textContent = message;
+  document.body.appendChild(alertContainer);
+};
+
+const showSuccessMsg = () => {
+  const message = successMsgTemplate.cloneNode(true);
+  showMessage(message, true);
+};
+
+const showErrorMsg = () => {
+  const message = errorMsgTemplate.cloneNode(true);
+  showMessage(message, false);
+};
+
 const isEscEvent = (evt) => evt.key === 'Escape' || evt.key === 'Esc';
 
 export {
@@ -37,5 +65,8 @@ export {
   getRandomArrayElement,
   getRandomArray,
   setFormEnabled,
+  showAlert,
+  showSuccessMsg,
+  showErrorMsg,
   isEscEvent
 };
